Copied fix from Fabric version
