--- src/main/java/com/mrh0/createaddition/blocks/portable_energy_interface/PEIActorVisual.java
@@ -1,35 +1,0 @@
-package com.mrh0.createaddition.blocks.portable_energy_interface;
-
-import com.simibubi.create.content.contraptions.actors.psi.PortableStorageInterfaceMovement;
-import com.simibubi.create.content.contraptions.behaviour.MovementContext;
-import com.simibubi.create.content.contraptions.render.ActorVisual;
-import com.simibubi.create.foundation.virtualWorld.VirtualRenderWorld;
-import dev.engine_room.flywheel.api.visualization.VisualizationContext;
-import net.createmod.catnip.animation.AnimationTickHolder;
-import net.createmod.catnip.animation.LerpedFloat;
-
-public class PEIActorVisual extends ActorVisual {
-
-	private final PortableEnergyInterfaceVisual.PIInstance instance;
-
-	public PEIActorVisual(VisualizationContext context, VirtualRenderWorld world, MovementContext movementContext) {
-		super(context, world, movementContext);
-
-		instance = new PortableEnergyInterfaceVisual.PIInstance(context.instancerProvider(), movementContext.state, movementContext.localPos, false);
-
-		instance.middle.light(localBlockLight(), 0);
-		instance.top.light(localBlockLight(), 0);
-	}
-
-	@Override
-	public void beginFrame() {
-		LerpedFloat lf = PortableStorageInterfaceMovement.getAnimation(context);
-		instance.tick(lf.settled());
-		instance.beginFrame(lf.getValue(AnimationTickHolder.getPartialTicks()));
-	}
-
-	@Override
-	protected void _delete() {
-		instance.remove();
-	}
-}
