+++ src/main/java/com/mrh0/createaddition/index/CABlockEntities.java
@@ -11,7 +11,6 @@ import com.mrh0.createaddition.blocks.electric_motor.*;
 import com.mrh0.createaddition.blocks.liquid_blaze_burner.*;
 import com.mrh0.createaddition.blocks.modular_accumulator.ModularAccumulatorBlockEntity;
 import com.mrh0.createaddition.blocks.modular_accumulator.ModularAccumulatorRenderer;
-import com.mrh0.createaddition.blocks.portable_energy_interface.PortableEnergyInterfaceVisual;
 import com.mrh0.createaddition.blocks.portable_energy_interface.PortableEnergyInterfaceRenderer;
 import com.mrh0.createaddition.blocks.portable_energy_interface.PortableEnergyInterfaceBlockEntity;
 import com.mrh0.createaddition.blocks.redstone_relay.RedstoneRelayBlockEntity;
